# Optimize bitmask finding for chunk size 1 and single chunk cases