1. With size-1 chunks, we needn't find unique labels in each chunk. There's only one label in each chunk!
2. When there is only a single chunk, we should just run `blockwise`.

```
| Before [497e7bc1] <main> | After [4935636b] | Ratio | Benchmark (Parameter)                       |
|--------------------------+------------------+-------+---------------------------------------------|
| 876±6μs                  | 640±80μs         |  0.73 | cohorts.SingleChunk.time_graph_construct    |
| 1.86±0.01ms              | 229±30μs         |  0.12 | cohorts.ERA5Google.time_find_group_cohorts  |
| 329±10μs                 | 1.58±0.1μs       |     0 | cohorts.SingleChunk.time_find_group_cohorts |

```